pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "dataagent-v2"
version = "0.1.0"
description = "Generic multi-step attribution agent — LangGraph + Vanna + LlamaIndex"
requires-python = ">=3.11"
dependencies = [
  # LangGraph engine
  "langgraph>=0.2",
  "langchain>=0.3",
  "langchain-openai>=0.2",
  "langchain-community>=0.3",
  # SQL layer
  "vanna[chromadb,openai]>=0.7",
  "sqlalchemy>=2.0",
  # Document RAG
  "llama-index-core>=0.11",
  "llama-index-readers-file>=0.2",
  "llama-index-vector-stores-chroma>=0.2",
  # Backend
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "python-dotenv>=1.0",
  "pyyaml>=6.0",
  "httpx>=0.27",
  # Utils
  "chromadb>=0.5",
  "pandas>=2.0",
  "networkx>=3.0",
]

[project.optional-dependencies]
dev = ["pytest", "ruff", "mypy"]
